refactor(ai-elements): split tool.tsx into per-tool modules
Split the monolithic 1834-line tool.tsx into focused per-tool module files
under a tool/ directory. This improves navigability and maintainability.

New structure:
- tool/ToolCollapsible.tsx: Shared expand/collapse pattern extracted from
  repeated usage across ~8 components
- tool/utils.ts: All utility functions, types, and constants
- tool/ToolIcon.tsx: Icon resolver component
- tool/ToolLabel.tsx: Label resolver component
- tool/ExecuteTool.tsx: Execute, ExecuteAuthRequired, WaitForExternalAuth
- tool/ReadFileTool.tsx: Read file viewer (now uses ToolCollapsible)
- tool/WriteFileTool.tsx: Write file diff viewer (now uses ToolCollapsible)
- tool/EditFilesTool.tsx: Edit files diff viewer (now uses ToolCollapsible)
- tool/CreateWorkspaceTool.tsx: Workspace creation with build logs
- tool/SubagentTool.tsx: Sub-agent status + SubagentStatusIcon
- tool/AgentReportTool.tsx: Collapsible markdown report
- tool/ChatSummarizedTool.tsx: Summary viewer (now uses ToolCollapsible)
- tool/Tool.tsx: Main dispatcher (memo-wrapped)
- tool/index.ts: Barrel re-exporting Tool, ToolIcon, ToolLabel

Existing imports from './tool' resolve to './tool/index.ts' automatically.
No behavioral changes.
