+++ site/src/components/ai-elements/tool/index.ts
@@ -1,0 +1,3 @@
+export { Tool } from "./Tool";
+export { ToolIcon } from "./ToolIcon";
+export { ToolLabel } from "./ToolLabel";
